requirements.txt
-e git+https://github.com/fernandoe/django-fe-core.git@release/0.4.0#egg=django_fe_core
-e git+https://github.com/fernandoe/django-fe-jwt.git@release/0.0.1#egg=django_fe_jwt
coveralls==1.5.1
dj-database-url==0.5.0
django-cors-headers==2.4.0
django-extensions==2.1.3
django-fe-version==0.0.1
Django==2.1.2
djangorestframework-jwt==1.11.0
djangorestframework==3.9.0
factory-boy==2.11.1
mysqlclient==1.3.13
pre-commit==1.11.2
pytest-cov==2.6.0
pytest-django==3.4.3

## The following requirements were added by pip freeze:
aspy.yaml==1.1.1
atomicwrites==1.2.1
attrs==18.2.0
cached-property==1.5.1
certifi==2018.10.15
cfgv==1.1.0
chardet==3.0.4
coverage==4.5.1
docopt==0.6.2
Faker==0.9.2
identify==1.1.7
idna==2.7
more-itertools==4.3.0
nodeenv==1.3.2
pluggy==0.8.0
py==1.7.0
PyJWT==1.6.4
pytest==3.9.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
text-unidecode==1.2
toml==0.10.0
urllib3==1.23
virtualenv==16.0.0
